Update
Update README and include some info about fine tuning.
