+++ train.py
@@ -49,7 +49,7 @@ from dataset import load_cleanunet_dataset
 from stft_loss import MultiResolutionSTFTLoss
 from util import rescale, find_max_epoch, print_size
 from util import LinearWarmupCosineDecay, loss_cleanunet
-
+from util import prepare_directories_and_logger, save_checkpoint, load_checkpoint
 from network import CleanUNet
 from logger import Logger
 
@@ -58,90 +58,7 @@ matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 plt.ioff()
 
-torch.autograd.set_detect_anomaly(True)
-
-# load checkpoint
-'''
-time0 = time.time()
-if log["ckpt_iter"] == 'max':
-    ckpt_iter = find_max_epoch(ckpt_directory)
-else:
-    ckpt_iter = log["ckpt_iter"]
-if ckpt_iter >= 0:
-    try:
-        # load checkpoint file
-        model_path = os.path.join(ckpt_directory, '{}.pkl'.format(ckpt_iter))
-        checkpoint = torch.load(model_path, map_location='cpu')
-        
-        # feed model dict and optimizer state
-        model.load_state_dict(checkpoint['model_state_dict'])
-        optimizer.load_state_dict(checkpoint['optimizer_state_dict'])
-
-        # record training time based on elapsed time
-        time0 -= checkpoint['training_time_seconds']
-        print('Model at iteration %s has been trained for %s seconds' % (ckpt_iter, checkpoint['training_time_seconds']))
-        print('checkpoint model loaded successfully')
-    except:
-        ckpt_iter = -1
-        print('No valid checkpoint model found, start training from initialization.')
-else:
-    ckpt_iter = -1
-    print('No valid checkpoint model found, start training from initialization.')
-'''
-def load_checkpoint(checkpoint_path, model):
-    assert os.path.isfile(checkpoint_path)
-    print("Loading checkpoint '{}'".format(checkpoint_path))
-    checkpoint_dict = torch.load(checkpoint_path, map_location='cpu')    
-    print(checkpoint_dict.keys())
-    model.load_state_dict(checkpoint_dict['model_state_dict'])
-    '''
-    missing_keys, unexpected_keys = model.load_state_dict(checkpoint_dict['state_dict'], strict=False)
-
-    if len(missing_keys) > 0:
-        print("Missing keys:", missing_keys)
-    if len(unexpected_keys) > 0:
-        print("Unexpected keys:", unexpected_keys)
-           
-    optimizer.load_state_dict(checkpoint_dict['optimizer'])
-    learning_rate = checkpoint_dict['learning_rate']
-    iteration = checkpoint_dict['iteration']
-    print("Loaded checkpoint '{}' from iteration {}" .format(
-        checkpoint_path, iteration))
-        
-    return model, optimizer, learning_rate, iteration
-    '''
-    return model
-
-
-def save_checkpoint(model, optimizer, learning_rate, iteration, filepath):
-    print("Saving model and optimizer state at iteration {} to {}".format(
-        iteration, filepath))
-    
-    torch.save({'iter': iteration,
-                'model_state_dict': model.state_dict(),
-                'optimizer_state_dict': optimizer.state_dict(),
-                'learning_rate': learning_rate}, 
-                filepath)    
-
-
-def prepare_directories_and_logger(output_dir, log_dir, ckpt_dir, rank):
-    if rank == 0:
-        if not os.path.isdir(output_dir):
-            os.makedirs(output_dir)
-            os.chmod(output_dir, 0o775)
-
-        if not os.path.isdir(log_dir):
-            os.makedirs(log_dir)
-            os.chmod(log_dir, 0o775)    
-
-        if not os.path.isdir(ckpt_dir):
-            os.makedirs(ckpt_dir)
-            os.chmod(ckpt_dir, 0o775)                        
-
-        logger = Logger(log_dir)
-    else:
-        logger = None
-    return logger
+#torch.autograd.set_detect_anomaly(True)
 
 
 def validate(model, val_loader, loss_fn, iteration, trainset_config, logger, rank, device):
